+++ src/co/insou/byteme/instruct/def/Equal_32.java
@@ -7,7 +7,7 @@ public class Equal_32 implements ByteMeInstruction {
 
     @Override
     public void execute(ByteMeRuntime runtime) {
-        runtime.stack().push((byte) (runtime.stack().pop() == runtime.stack().pop() ? 1 : 0));
+        runtime.stack().push(runtime.stack().pop() == runtime.stack().pop() ? 1 : 0);
     }
 
 }
